+++ ffi/dnp3-schema/Cargo.toml
@@ -7,3 +7,4 @@ edition = "2021"
 
 [dependencies]
 oo-bindgen = { git = "https://github.com/stepfunc/oo_bindgen.git", tag = "0.3.0" }
+tracing-ffi-schema = { git = "https://github.com/stepfunc/tracing-ffi.git" }
