+++ ffi/dnp3-schema/src/decoding.rs
@@ -1,0 +1,79 @@
+use oo_bindgen::model::*;
+
+const NOTHING: &str = "nothing";
+
+pub fn define(lib: &mut LibraryBuilder) -> BackTraced<UniversalStructHandle> {
+    let app_decode_level_enum = lib
+        .define_enum("app_decode_level")?
+        .push(NOTHING, "Decode nothing")?
+        .push("header", " Decode the header-only")?
+        .push("object_headers", "Decode the header and the object headers")?
+        .push(
+            "object_values",
+            "Decode the header, the object headers, and the object values",
+        )?
+        .doc("Controls how transmitted and received application-layer fragments are decoded at the INFO log level")?
+        .build()?;
+
+    let transport_decode_level_enum = lib
+        .define_enum("transport_decode_level")?
+        .push(NOTHING, "Decode nothing")?
+        .push("header", " Decode the header")?
+        .push("payload", "Decode the header and the raw payload as hexadecimal")?
+        .doc("Controls how transmitted and received transport segments are decoded at the INFO log level")?
+        .build()?;
+
+    let link_decode_level_enum = lib
+        .define_enum("link_decode_level")?
+        .push(NOTHING, "Decode nothing")?
+        .push("header", " Decode the header")?
+        .push(
+            "payload",
+            "Decode the header and the raw payload as hexadecimal",
+        )?
+        .doc("Controls how transmitted and received link frames are decoded at the INFO log level")?
+        .build()?;
+
+    let phys_decode_level_enum = lib
+        .define_enum("phys_decode_level")?
+        .push(NOTHING, "Log nothing")?
+        .push(
+            "length",
+            "Log only the length of data that is sent and received",
+        )?
+        .push(
+            "data",
+            "Log the length and the actual data that is sent and received",
+        )?
+        .doc("Controls how data transmitted at the physical layer (TCP, serial, etc) is logged")?
+        .build()?;
+
+    let application_field = Name::create("application")?;
+    let transport_field = Name::create("transport")?;
+    let link_field = Name::create("link")?;
+    let physical_field = Name::create("physical")?;
+
+    let decode_level_struct = lib.declare_universal_struct("decode_level")?;
+    let decode_level_struct = lib.define_universal_struct(decode_level_struct)?
+        .add(&application_field, app_decode_level_enum, "Controls application fragment decoding")?
+        .add(&transport_field, transport_decode_level_enum, "Controls transport segment layer decoding")?
+        .add(&link_field, link_decode_level_enum, "Controls link frame decoding")?
+        .add(&physical_field, phys_decode_level_enum, "Controls the logging of physical layer read/write")?
+        .doc("Controls the decoding of transmitted and received data at the application, transport, link, and physical layers")?
+        .end_fields()?
+        .begin_initializer("init", InitializerType::Normal, "Initialize log levels to defaults")?
+        .default_variant(&application_field, NOTHING)?
+        .default_variant(&transport_field, NOTHING)?
+        .default_variant(&link_field, NOTHING)?
+        .default_variant(&physical_field, NOTHING)?
+        .end_initializer()?
+        .begin_initializer("nothing", InitializerType::Static, "Initialize log levels to nothing")?
+        .default_variant(&application_field, NOTHING)?
+        .default_variant(&transport_field, NOTHING)?
+        .default_variant(&link_field, NOTHING)?
+        .default_variant(&physical_field, NOTHING)?
+        .end_initializer()?
+        .build()?;
+
+    Ok(decode_level_struct)
+}
